requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
jsonschema==2.6.0
pendulum==1.2.0
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
requests==2.24.0
simplejson==3.11.1
singer-python==2.1.4
six==1.15.0
target-csv==0.3.0
tzlocal==2.1
urllib3==1.25.10
